Redeploy
