+++ webpack/connectivity/__tests__/refetcher_test.ts
@@ -1,0 +1,5 @@
+describe("startRefetcher", () => {
+  it("detects network edge changes", () => {
+    fail();
+  });
+});
